gui/requirements.txt
pycryptodome
pyperclip
bcrypt